+++ src/main/java/com/atlauncher/utils/ImportPackUtils.java
@@ -28,6 +28,7 @@ import java.util.Properties;
 import com.atlauncher.FileSystem;
 import com.atlauncher.Gsons;
 import com.atlauncher.data.curseforge.pack.CurseForgeManifest;
+import com.atlauncher.data.modrinth.pack.ModrinthModpackManifest;
 import com.atlauncher.data.multimc.MultiMCInstanceConfig;
 import com.atlauncher.data.multimc.MultiMCManifest;
 import com.atlauncher.gui.dialogs.InstanceInstallerDialog;
@@ -54,6 +55,10 @@ public class ImportPackUtils {
         try {
             if (ArchiveUtils.archiveContainsFile(file.toPath(), "manifest.json")) {
                 return loadCurseForgeFormat(file, null, null);
+            }
+
+            if (ArchiveUtils.archiveContainsFile(file.toPath(), "index.json")) {
+                return loadModrinthFormat(file);
             }
 
             Path tmpDir = FileSystem.TEMP.resolve("multimcimport" + file.getName().toString().toLowerCase());
@@ -109,6 +114,45 @@ public class ImportPackUtils {
             new InstanceInstallerDialog(manifest, tmpDir);
         } catch (Exception e) {
             LogManager.logStackTrace("Failed to install CurseForge pack", e);
+            FileUtils.deleteDirectory(tmpDir);
+            return false;
+        }
+
+        return true;
+    }
+
+    public static boolean loadModrinthFormat(File file) {
+        if (!file.getName().endsWith(".zip")) {
+            LogManager.error("Cannot install as the file was not a zip file");
+            return false;
+        }
+
+        Path tmpDir = FileSystem.TEMP.resolve("modrinthimport" + file.getName().toString().toLowerCase());
+
+        try {
+            ModrinthModpackManifest manifest = Gsons.MINECRAFT
+                    .fromJson(ArchiveUtils.getFile(file.toPath(), "index.json"), ModrinthModpackManifest.class);
+
+            if (!manifest.game.equals("minecraft")) {
+                LogManager.error(
+                        "Cannot install as the manifest is for game " + manifest.game + " and not for Minecraft");
+                return false;
+            }
+
+            if (!manifest.dependencies.containsKey("minecraft")) {
+                LogManager.error("Cannot install as the manifest doesn't contain a minecraft dependency");
+                return false;
+            }
+
+            if (Integer.parseInt(manifest.formatVersion, 10) != 1) {
+                LogManager.warn("Manifest is version " + manifest.formatVersion + " which may be an issue!");
+            }
+
+            ArchiveUtils.extract(file.toPath(), tmpDir);
+
+            new InstanceInstallerDialog(manifest, tmpDir);
+        } catch (Exception e) {
+            LogManager.logStackTrace("Failed to install Modrinth pack", e);
             FileUtils.deleteDirectory(tmpDir);
             return false;
         }